ci: run on push, use LLVM/clang 18 (FIR-42481) (#70)
FIR-42481: [M1] Upgrade requirements for Knight Bus (Arrow & friends)

Add a post-merge CI, mostly so that the ccache is populated for branch
pulls. I originally removed this because it caused the CI to be run on
PRs twice, but instead I now changed which branches this applies to.

Also upgrade the LLVM/clang version used in the CI to 18, and the Ubuntu
version used for sanitized builds from 22.04 to 24.04, as the version of
boost shipped with 22.04 is too old to compile with clang-18. This isn't
an issue for the non-sanitized builds because it installs dependencies
with conda instead of apt.