fix: ParseMetadata: handle case when corded buffer does not start at offset 0 (FIR-42560) (#71)
FIR-42560: [M1] Basic Parquet Reader based on Apache Arrow + Buffer
Manager

Fix the "happy path" of parquet footer parsing, when the entire footer
is contained in a single slice. The issue was that we assumed that the
slice started with the beginning of the footer, but this is not true: it
might contain arbitrary data *before* the footer as well if we fetched
too much.

This was previously masked by a bug in our parquet reader, fixed in
https://github.com/firebolt-analytics/packdb/pull/16042