fix(arrow): prevent unaligned buffers (FIR-47668) (#75)
Sometimes ReadBuffer might return a Splice() of the raw socket buffer,
which might be unaligned with the type it actually holds.
This causes errors in UBsan and also might cause SIGSEGV in old
compilers even on x86.
To prevent this, we create 64-byte aligned copy of the buffer if it is
not 8-byte aligned.

> On most modern CPUs this is not an issue, but some older CPUs may
crash or suffer poor performance. For this reason it is recommended that
all incoming array buffers are properly aligned, but some data sources
such as
[Flight](https://arrow.apache.org/docs/format/Flight.html#flight-rpc)
may produce unaligned buffers.

Source: https://arrow.apache.org/docs/cpp/env_vars.html