fix(arrow): allow dictionary-encoded booleans (FIR-49299) (#78)
We need to be able to read an Iceberg table created by Snowflake.
This file contains a dictionary-encoded boolean column. (This is
probably a mistake on Snowflake's end.) Since there's not a good reason
to have such a column, Arrow currently blocks when trying to read it.

Remove lines which explicitly block reading from dictionary-encoded
boolean.

Manual test for now. Will add SQL test when updating PackDB to point to
this commit.

Without fix:
```
⚡︎select * from read_parquet('s3://.../parquet/dict_encoded_bool.parquet');
Error reading column 'a' in row group 0 of 's3://.../parquet/dict_encoded_bool.parquet': IOError: Not yet implemented: Dictionary encoding not implemented for boolean type.
```
With fix:
```
⚡︎select * from read_parquet('s3://.../parquet/dict_encoded_bool.parquet');
┌─a─┐
│ t │
└───┘
```